# docs/integrations/api/openapi.yaml
openapi: 3.0.3
info:
  title: Landslide Monitoring API (v2)
  version: 0.1.0
  description: |
    本文件是 API 契约的机器可读入口（契约优先），与 `docs/integrations/api/*.md` 保持一致。
servers:
  - url: /api/v1
...
security:
  - BearerAuth: []
paths:
  /auth/login:
    post:
      tags: [auth]
      summary: User login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/LoginResponse"
        "400":
          $ref: "#/components/responses/ErrorResponse"
        "500":
          $ref: "#/components/responses/InternalServerErrorResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
  /auth/refresh:
    post:
      tags: [auth]
      summary: Refresh token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshRequest"
      responses:
        "200":
          $ref: "#/components/responses/SuccessResponse"
        "400":
          $ref: "#/components/responses/ErrorResponse"
        "500":
          $ref: "#/components/responses/InternalServerErrorResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    SuccessResponse:
      description: ok
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/SuccessResponse"
    ErrorResponse:
      description: error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    InternalServerErrorResponse:
      description: internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    SuccessResponse:
      type: object
      required: [success, code, message, timestamp, traceId]
      properties:
        success: { type: boolean, enum: [true] }
        code: { type: integer, example: 200 }
        message: { type: string, example: ok }
        data: {}
        timestamp:
          type: string
          format: date-time
          example: "2025-12-15T10:00:00Z"
        traceId:
          type: string
          example: "req_01J..."
    ErrorResponse:
      type: object
      required: [success, code, message, timestamp, traceId]
      properties:
        success: { type: boolean, enum: [false] }
        code: { type: integer, example: 400 }
        message: { type: string, example: 参数错误 }
        error:
          type: object
          additionalProperties: true
        timestamp:
          type: string
          format: date-time
          example: "2025-12-15T10:00:00Z"
        traceId:
          type: string
          example: "req_01J..."
...
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username: { type: string }
        password: { type: string }
    RefreshRequest:
      type: object
      required: [refreshToken]
      properties:
        refreshToken: { type: string }
...
    LoginResponse:
      type: object
      required: [token, refreshToken, expiresIn, user]
      properties:
        token: { type: string }
        refreshToken: { type: string }
        expiresIn: { type: integer }
        user:
          type: object
          required: [userId, username, roles, permissions]
          properties:
            userId: { type: string, format: uuid }
            username: { type: string }
            realName: { type: string }
            roles:
              type: array
              items: { type: string }
            permissions:
              type: array
              items: { type: string }
...